Refactor screen files
