+++ lib/main.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:shop_flutter_app/dependencies.dart';
+import 'package:shop_flutter_app/screens/main/main_screen.dart';
 import 'package:shop_flutter_app/screens/navigator.dart';
-import 'package:shop_flutter_app/screens/cart_tab.dart';
-import 'package:shop_flutter_app/screens/main_tab.dart';
-import 'package:shop_flutter_app/screens/profile_tab.dart';
 
 void main() async {
   await Dependencies.init();
@@ -34,159 +32,9 @@ class MyApp extends StatelessWidget {
           // is not restarted.
           primarySwatch: Colors.blue,
         ),
-        home: MainBottomNavigationBar(key: AppNavigator.bottomBarKey),
+        home: MainScreen(key: AppNavigator.bottomBarKey),
         navigatorKey: AppNavigator.navigatorKey,
       ),
-    );
-  }
-}
-
-class MainBottomNavigationBar extends StatefulWidget {
-  const MainBottomNavigationBar({Key? key}) : super(key: key);
-
-  @override
-  State<MainBottomNavigationBar> createState() =>
-      MainBottomNavigationBarState();
-}
-
-class MainBottomNavigationBarState extends State<MainBottomNavigationBar> {
-  int currentTab = 0;
-
-  openMainPage() => setState(() => currentTab = 0);
-  openCartPage() => setState(() => currentTab = 1);
-  openProfilePage() => setState(() => currentTab = 2);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      // appBar: AppBar(
-      //   title: const Text('TabBar Widget'),
-      // ),
-      body: AnimatedSwitcher(
-        duration: Duration(milliseconds: 200),
-        child: <Widget>[
-          const MainTab(),
-          const CartTab(),
-          const ProfileTab(),
-        ][currentTab],
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: currentTab,
-        onTap: (int i) {
-          setState(() => currentTab = i);
-        },
-        items: [
-          BottomNavigationBarItem(
-            icon: Container(
-              width: 10,
-              height: 10,
-              color: Colors.red,
-            ),
-            label: 'Main',
-          ),
-          BottomNavigationBarItem(
-            icon: Container(
-              width: 10,
-              height: 10,
-              color: Colors.red,
-            ),
-            label: 'Cart',
-          ),
-          BottomNavigationBarItem(
-            icon: Container(
-              width: 10,
-              height: 10,
-              color: Colors.red,
-            ),
-            label: 'Profile',
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key, required this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headline4,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
